# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tumbas POS API",
    "description": "Minimal point of sale API for products, checkout, sales history, and sales summary.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
...
  "paths": {
...
    "/api/v1/inventory/receiving/{id}": {
      "get": {
        "tags": [
          "Receiving"
        ],
        "summary": "Show receiving detail",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Receiving details."
          }
        }
      },
      "put": {
        "tags": [
          "Receiving"
        ],
        "summary": "Update draft receiving / Finalize",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Receiving updated."
          }
        }
      },
      "delete": {
        "tags": [
          "Receiving"
        ],
        "summary": "Delete draft receiving",
        "description": "Deletes a draft stock receiving. Completed receiving cannot be deleted. Restricted to manage_inventory permission.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Draft receiving deleted."
          }
        }
      }
    },
...
    "/api/v1/inventory/suppliers/{id}": {
      "get": {
        "tags": [
          "Suppliers"
        ],
        "summary": "Show supplier details",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier details."
          }
        }
      },
      "put": {
        "tags": [
          "Suppliers"
        ],
        "summary": "Update supplier details",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nama": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "nullable": true
                  },
                  "nomor_telepon": {
                    "type": "string",
                    "nullable": true
                  },
                  "alamat": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": [
                  "nama"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Supplier updated."
          }
        }
      },
      "delete": {
        "tags": [
          "Suppliers"
        ],
        "summary": "Delete supplier master",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier deleted."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the Sanctum plainTextToken retrieved from login."
      }
    },
...
  }
}